Implemented task 7
One get request is to create the playlist table another is to get the trackIDs once they are in a playlist. A function was created to send the data to mysql. Finally a post request using the function was created to get user input as an input to the function.